requirements.txt
streamlit==1.39.0
langchain==0.3.4
openai==1.52.0
python-dotenv==1.0.1
langchain_community==0.3.3
watchdog==5.0.3
langchain_openai==0.2.3